minimax refactor (#2)

+++ src/uci/commands.rs
@@ -34,8 +34,8 @@ pub enum UciOutput {
 
 #[derive(Debug)]
 pub struct Info {
-    pub depth: u32,
-    pub sel_depth: u32,
+    pub depth: u64,
+    pub sel_depth: u64,
     pub nodes: u32,
     pub nodes_per_second: u32,
     pub time: u32,
